Add worker.Fetcher

+++ storage/record.go
@@ -3,7 +3,7 @@ package storage
 type Record struct {
 	Id       int            `json:"id"`
 	Url      string         `json:"url"`
-	Interval int            `json:"interval"`
+	Interval float64        `json:"interval"`
 	History  []HistoryEntry `json:"-"`
 }
